requirements.txt
grpcio==1.54.2
grpcio-tools==1.54.2
requests==2.31.0